financial-model/requirements.txt
flask>=3.0.0
numpy-financial>=1.0.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
